Refactor LanguageDropdownSelector
